new ARnftThreejsComp component
